[MON-129] feat:  SeriesAssemble Layer 이용해서 코드 리팩토링 - SeriesAssemble 를 만들었습니다. - 시리즈 공고 발행시 사진없는 상태에서 시리즈 저장하고, 이미지는 업데이트 하는 방식으로 변경하였습니다.
