+++ src/main/java/com/prgrms/monthsub/module/series/series/converter/SeriesConverter.java
@@ -48,11 +48,9 @@ public class SeriesConverter {
 
   public Series SeriesSubscribePostResponseToEntity(
     Writer writer,
-    String imageUrl,
     SeriesSubscribePost.Request req
   ) {
     return Series.builder()
-      .thumbnailKey(imageUrl)
       .title(req.title())
       .introduceText(req.introduceText())
       .introduceSentence(req.introduceSentence())
@@ -216,5 +214,3 @@ public class SeriesConverter {
   }
 
 }
-
-
